+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v3.0.1 (upcomming)
 - added the link to excel upload again
 - now showing a no character instead of null on student search
 - removed location from student search screen
+- added the log screen to show all that has been logged
 
 V3.0.0 (16-11-2018)
 - Redesigned the frontend